tests: refactor unit tests to match new layout
